feat(hub-common): separate util tests
